refactor(daemon): group data-migration modules into migration/
