+++ apps/daemon/src/migration/legacy-data-migrator.ts
